refactor: helpers files and folder structure
